feature/mini-garbage collector's done
