+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: tde-souz <tde-souz@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/20 19:50:29 by thfirmin          #+#    #+#             */
-/*   Updated: 2023/04/29 09:43:31 by thfirmin         ###   ########.fr       */
+/*   Updated: 2023/05/12 14:53:16 by thfirmin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,12 +17,12 @@
 # include "includes/ft_printf.h"
 
 // Struct
-typedef struct s_list
+typedef struct s_mem
 {
-	void			*content;
-	int				value;
-	struct s_list	*next;
-}					t_list;
+	void			*ptr;
+	char			*context;
+	struct s_mem	*next;
+}					t_mem;
 
 // is
 int		ft_isalpha(int c);
@@ -61,7 +61,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len);
 void	*ft_memchr(const void *s, int c, size_t n);
 int		ft_memcmp(const void *s1, const void *s2, size_t n);
 void	ft_bzero(void *s, size_t n);
-void	*ft_calloc(size_t count, size_t size);
+void    *ft_calloc(size_t count, size_t size, char *context, t_mem **mem);
 
 // Put
 int		ft_putchar_fd(char c, int fd);
@@ -69,15 +69,15 @@ int		ft_putstr_fd(char *s, int fd);
 int		ft_putendl_fd(char *s, int fd);
 int		ft_putnbr_fd(int n, int fd);
 
-// Bonus
-void	ft_lstadd_back(t_list **lst, t_list *new);
-void	ft_lstadd_front(t_list **lst, t_list *new);
-void	ft_lstclear(t_list **lst, void (*del)(void *));
-void	ft_lstdelone(t_list *lst, void (*del)(void *));
-void	ft_lstiter(t_list *lst, void (*f)(void *));
-t_list	*ft_lstlast(t_list *lst);
-t_list	*ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
-t_list	*ft_lstnew(void *content, int value);
-int		ft_lstsize(t_list *lst);
+// Garbage Collector
+void	ft_memadd_back(t_mem **mem, t_mem *new);
+void	ft_memadd_front(t_mem **mem, t_mem *new);
+t_mem	*ft_memnew(void	*ptr, char *context);
+void	*ft_malloc(size_t size, char *context, t_mem **mem);
+void	*ft_calloc(size_t count, size_t size, char *context, t_mem **mem);
+void	ft_free(void *ptr, t_mem **mem);
+void	ft_freel(t_mem **mem);
+void	ft_sumary(t_mem *mem);
+int		ft_memsize(t_mem *mem);
 
 #endif
