+++ libft/srcs/ft/ft_freel.c
@@ -6,7 +6,7 @@
 /*   By: thfirmin <thfirmin@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/12 14:12:16 by thfirmin          #+#    #+#             */
-/*   Updated: 2023/05/12 14:12:40 by thfirmin         ###   ########.fr       */
+/*   Updated: 2023/05/13 13:20:10 by thfirmin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,16 @@
 
 void	ft_freel(t_mem **mem)
 {
-	(void) mem;
+	t_mem	*nxt;
+
+	if (!mem)
+		return ;
+	while (*mem)
+	{
+		nxt = (**mem).next;
+		free ((**mem).ptr);
+		free ((**mem).context);
+		free (*mem);
+		*mem = nxt;
+	}
 }
