--- libft/srcs/ft/ft_lstdelone_bonus.c
@@ -1,21 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ft_lstdelone_bonus.c                               :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: thfirmin <thfirmin@student.42.rio>         +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/06 15:02:21 by thfirmin          #+#    #+#             */
-/*   Updated: 2023/01/20 23:48:26 by thfirmin         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "libft.h"
-
-// Delete one gived node
-void	ft_lstdelone(t_list *lst, void (*del)(void *))
-{
-	if (del)
-		del(lst->content);
-	free (lst);
-}
