+++ libft/srcs/ft/ft_sumary.c
@@ -6,7 +6,7 @@
 /*   By: thfirmin <thfirmin@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/12 14:13:50 by thfirmin          #+#    #+#             */
-/*   Updated: 2023/05/12 14:14:08 by thfirmin         ###   ########.fr       */
+/*   Updated: 2023/05/14 10:12:53 by thfirmin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,18 @@
 
 void	ft_sumary(t_mem *mem)
 {
-	(void) mem;
+	int	i;
+
+	ft_printf ("\n--------------------| SUMARY |-------------------\n");
+	ft_printf ("Bytes allocate'd: %d\n", ft_memsize(mem));
+	i = -1;
+	while (mem)
+	{
+		ft_printf ("----------------------------------------\n");
+		ft_printf ("block[%d]:\t", ++i);
+		ft_printf ("%p ", mem->ptr);
+		ft_printf ("[%s]\n", mem->context);
+		mem = mem->next;
+	}
+	ft_printf ("--------------------| END |-------------------\n");
 }
